Add new icons
