--- app/src/main/java/mattecarra/accapp/fragments/ExportViewModel.kt
@@ -1,14 +1,0 @@
-package mattecarra.accapp.fragments
-
-import android.app.Application
-
-import androidx.lifecycle.AndroidViewModel
-import androidx.lifecycle.LiveData
-import mattecarra.accapp.database.AccaRoomDatabase
-import mattecarra.accapp.database.ProfileDao
-import mattecarra.accapp.models.AccaProfile
-
-class ExportViewModel(application: Application) : AndroidViewModel(application) {
-
-
-}
