+++ app/src/main/res/drawable/ic_export.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M12,1L8,5h3v9h2V5h3m2,18H6a2,2 0,0 1,-2 -2V9a2,2 0,0 1,2 -2h3v2H6v12h12V9h-3V7h3a2,2 0,0 1,2 2v12a2,2 0,0 1,-2 2z"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M0,0h24v24h-24z"
+      android:fillColor="#00000000"/>
+</vector>
