--- app/src/main/res/layout/activity_import_export.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        android:fitsSystemWindows="true"
-        tools:context=".activities.ImportExportActivity">
-
-    <com.google.android.material.appbar.AppBarLayout
-            android:id="@+id/import_export_appbarlayout"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-
-        <com.google.android.material.appbar.MaterialToolbar
-                android:id="@+id/import_export_toolbar"
-                style="?toolbarStyle"
-                android:layout_width="match_parent"
-                android:layout_height="?attr/actionBarSize"
-                app:layout_scrollFlags="enterAlways|scroll"
-                app:title="@string/import_export_title" />
-
-        <com.google.android.material.tabs.TabLayout
-                android:id="@+id/import_export_tabs"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                app:tabGravity="fill"
-                app:tabMode="fixed"/>
-    </com.google.android.material.appbar.AppBarLayout>
-
-    <androidx.viewpager2.widget.ViewPager2
-            android:id="@+id/import_export_viewpager"
-            android:layout_below="@+id/import_export_appbarlayout"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent" />
-
-    <com.google.android.material.floatingactionbutton.ExtendedFloatingActionButton
-            android:id="@+id/import_export_fab"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentBottom="true"
-            android:layout_centerHorizontal="true"
-            android:layout_margin="16dp"
-            android:text="@string/import_title"
-            app:icon="@drawable/ic_outline_save_alt_24px" />
-
-</RelativeLayout>
