--- app/src/main/res/layout/fragment_export.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context=".fragments.ExportFragment">
-
-    <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/export_list_rv"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:foreground="?android:selectableItemBackground"
-            android:scrollbars="vertical"
-            tools:listitem="@layout/profile_export_item"/>
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
