--- app/src/main/res/layout/fragment_import.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context=".fragments.ImportFragment">
-
-    <Button
-            android:id="@+id/import_load_clipboard_btn"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_centerHorizontal="true"
-            android:drawableStart="@drawable/ic_save"
-            android:drawableLeft="@drawable/ic_save"
-            android:text="@string/import_load_from_clipboard" />
-    
-    <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/import_profiles_rv"
-            android:layout_below="@+id/import_load_clipboard_btn"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:visibility="gone"/>
-
-    <TextView
-            android:id="@+id/import_profile_empty_tv"
-            android:layout_centerInParent="true"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Press Load Profiles from Clipboard!"/>
-
-</RelativeLayout>
